backend/backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
